+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/EntityViewManagerImpl.java
@@ -919,16 +919,16 @@ public class EntityViewManagerImpl implements EntityViewManager {
     public String applyObjectBuilder(ManagedViewTypeImplementor<?> viewType, MappingConstructorImpl<?> mappingConstructor, String entityViewRoot, FullQueryBuilder<?, ?> criteriaBuilder, EntityViewConfiguration configuration, int offset) {
         Path root = getPath(criteriaBuilder, entityViewRoot);
         String path = root.getPath();
-        criteriaBuilder.selectNew(createObjectBuilder(viewType, mappingConstructor, root.getJavaType(), path, null, criteriaBuilder, configuration, offset, 0));
+        criteriaBuilder.selectNew(createObjectBuilder(viewType, mappingConstructor, root.getJavaType(), path, null, criteriaBuilder, configuration, offset, 0, false));
         return path;
     }
 
-    public ObjectBuilder<?> createObjectBuilder(ManagedViewTypeImplementor<?> viewType, MappingConstructorImpl<?> mappingConstructor, String entityViewRoot, String embeddingViewPath, FullQueryBuilder<?, ?> criteriaBuilder, EntityViewConfiguration configuration, int offset, int suffix) {
+    public ObjectBuilder<?> createObjectBuilder(ManagedViewTypeImplementor<?> viewType, MappingConstructorImpl<?> mappingConstructor, String entityViewRoot, String embeddingViewPath, FullQueryBuilder<?, ?> criteriaBuilder, EntityViewConfiguration configuration, int offset, int suffix, boolean nullFlatViewIfEmpty) {
         Path root = getPath(criteriaBuilder, entityViewRoot);
-        return createObjectBuilder(viewType, mappingConstructor, root.getJavaType(), root.getPath(), embeddingViewPath, criteriaBuilder, configuration, offset, suffix);
+        return createObjectBuilder(viewType, mappingConstructor, root.getJavaType(), root.getPath(), embeddingViewPath, criteriaBuilder, configuration, offset, suffix, nullFlatViewIfEmpty);
     }
 
-    public ObjectBuilder<?> createObjectBuilder(ManagedViewTypeImplementor<?> viewType, MappingConstructorImpl<?> mappingConstructor, Class<?> rootType, String entityViewRoot, String embeddingViewPath, FullQueryBuilder<?, ?> criteriaBuilder, EntityViewConfiguration configuration, int offset, int suffix) {
+    public ObjectBuilder<?> createObjectBuilder(ManagedViewTypeImplementor<?> viewType, MappingConstructorImpl<?> mappingConstructor, Class<?> rootType, String entityViewRoot, String embeddingViewPath, FullQueryBuilder<?, ?> criteriaBuilder, EntityViewConfiguration configuration, int offset, int suffix, boolean nullFlatViewIfEmpty) {
         ExpressionFactory ef = criteriaBuilder.getService(ExpressionFactory.class);
         if (!viewType.getEntityClass().isAssignableFrom(rootType)) {
             if (rootType.isAssignableFrom(viewType.getEntityClass())) {
@@ -951,7 +951,7 @@ public class EntityViewManagerImpl implements EntityViewManager {
         criteriaBuilder.registerMacro("view_root", viewRootJpqlMacro);
 
         return getTemplate(macroEf, viewType, mappingConstructor, entityViewRoot, embeddingViewPath, embeddingViewJpqlMacro, offset)
-            .createObjectBuilder(criteriaBuilder, configuration.getOptionalParameters(), configuration, suffix);
+            .createObjectBuilder(criteriaBuilder, configuration.getOptionalParameters(), configuration, suffix, false, nullFlatViewIfEmpty);
     }
 
     private static Path getPath(FullQueryBuilder<?, ?> queryBuilder, String entityViewRoot) {