+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/update/EntityViewUpdaterImpl.java
@@ -169,7 +169,7 @@ public class EntityViewUpdaterImpl implements EntityViewUpdater {
                             null,
                             new MutableEmbeddingViewJpqlMacro(),
                             0
-                    ).createObjectBuilder(null, null, null, 0);
+                    ).createObjectBuilder(null, null, null, 0, false, false);
                     jpaIdInstantiator = null;
                 } else {
                     tupleizer = null;