+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/correlation/subselectsubset/model/DocumentSubselectElementCollectionView.java
@@ -17,20 +17,29 @@
 package com.blazebit.persistence.view.testsuite.correlation.subselectsubset.model;
 
 import com.blazebit.persistence.testsuite.entity.Document;
+import com.blazebit.persistence.testsuite.entity.Person;
 import com.blazebit.persistence.view.EntityView;
+import com.blazebit.persistence.view.FetchStrategy;
 import com.blazebit.persistence.view.IdMapping;
+import com.blazebit.persistence.view.Mapping;
+import com.blazebit.persistence.view.MappingCorrelatedSimple;
+
+import java.util.Set;
 
 /**
  *
  * @author Christian Beikov
- * @since 1.4.0
+ * @since 1.4.1
  */
 @EntityView(Document.class)
-public interface DocumentSubselectView {
-    
+public interface DocumentSubselectElementCollectionView {
+
     @IdMapping
     public Long getId();
 
     public String getName();
+
+    @Mapping(fetch = FetchStrategy.SUBSELECT)
+    public Set<NameObjectView> getNames();
 
 }